course_repo/pyproject.toml
[tool.poetry]
name = "rl-foundations-for-llms-harness"
version = "0.1.0"
description = "Minimal, local, verifiable-reward harness for an RL foundations course."
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
groq = ">=0.9.0"
python-dotenv = ">=1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0"
matplotlib = ">=3.7"
tiktoken = ">=0.5.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
